Fixing path for images in generated rdoc
In the last commit I change the README.rdoc to show images in
github page. Thus I insert the path for image files in github. The
problem is that it brokes the docs generated by rdoc command in a
local directory. I fix that but in a bad way, now we need the next
lines to generate de rdoc:

$ rake index_file
$ rake doc

In a future I like only:

$ rake rdoc